Day 19.2: Now with well-polished input lists.
Hopefully.
Takes a few minutes for each x, times some 300,
that comes up to a day or so...

Day 17.1:
Also some futzing with 17.2,
using known too high numbers as the minimum.
Still looks evil.
Perhaps make it find the suboptimal,
known bad route first,
then apply the run with leeway.
Also, leeway may probably consist of number of extra steps to take?
Which is not particularly well-known :(

Changes to be committed:
	modified:   ../Day17/Day17.1.py
	modified:   Day19.2.py